+++ src/server/root/src/root.rs
@@ -124,7 +124,7 @@ fn alloc_mem(is: &mut GateIStream, child: &mut dyn Child) {
     let dst_sel: Selector = is.pop();
     let addr: goff = is.pop();
     let size: usize = is.pop();
-    let perms = kif::Perm::from_bits_truncate(is.pop::<u8>());
+    let perms = kif::Perm::from_bits_truncate(is.pop::<u32>());
 
     let res = if addr == !0 {
         memory::get().allocate_for(child, dst_sel, size, perms)
